seperate test in different components
