+++ .babelrc
@@ -1,4 +1,4 @@
 {
   "presets": ["@babel/preset-react"],
-  "plugins": ["@babel/plugin-syntax-jsx"]
+  "plugins": ["@babel/plugin-syntax-jsx", "react-data-testid"]
 }
