+++ package.json
@@ -45,6 +45,7 @@
     "@babel/preset-env": "^7.18.9",
     "@babel/preset-react": "^7.18.6",
     "@testing-library/react": "^13.3.0",
+    "babel-plugin-react-data-testid": "^0.2.0",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-plugin-import": "^2.26.0",
