+++ src/components/Quote.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import Quote from './Quote';
+
+it('renders correctly', () => {
+  const tree = renderer.create(<Quote />).toJSON();
+  expect(tree).toMatchSnapshot();
+});
