+++ src/logic/calculate.test.js
@@ -5,4 +5,14 @@ describe('calculate', () => {
     const result = calculate({ total: 0, next: null, operation: null }, 'AC');
     expect(result).toEqual({ total: 0, next: null, operation: null });
   });
+
+  it('test "-" button', () => {
+    const obj = {
+      total: '100',
+      next: '4',
+      operation: '-',
+    };
+    const result = calculate(obj, '=');
+    expect(result).toEqual({ total: '96', next: null, operation: null });
+  });
 });
